pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "great-humans"
version = "1.0.0"
description = "Python library for the 100 Great Humans dataset"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Great Humans Project"}
]
keywords = ["history", "agents", "ai", "personas", "dataset"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pyyaml>=6.0.1",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.urls]
Homepage = "https://github.com/Gzeu/100-great-humans"
Repository = "https://github.com/Gzeu/100-great-humans"
Issues = "https://github.com/Gzeu/100-great-humans/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
